requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
email_validator==2.1.0
gunicorn==21.2.0